[BE] reco zzim fix

+++ BE/zeepseek/src/main/java/com/zeepseek/backend/domain/recommend/dto/response/RecommendationDto.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class RecommendationDto {
-    private Long propertyId;
+    private int propertyId;
     private double similarity;
 }
